refactoring(tui): replace generic state with typed state with observers
Replacing generic state that I created before with new typed state
specific for current app is better in multiple ways.
The code is more readable now, and the data flow and reaction to them is
easier to work with. I feel that ListLog is much easier to maintain.
I've also fixed all bugs regarding scrolling. There is still one more
with regenerating number of prerendered components in list. Will fix
that in next commit.
I'm also wondering if maybe scrolling over all lines, not only entire
items could be possible and a little bit better. But not at this moment.