fix(rendering): optimize idle cpu usage
At the beginning I thought that php-tui will handle rerendering
optimisation, but unfortunately it's not, and I don't know if it's event
possible.
Because I have global state, I'm able to observe it and rerender view
only when there is a change in it. The only important thing is to
remember that everything that has to rerender view, has to be in state.
It's less flexible and also it's less snappy now, but it's reducing CPU
usage in idle from above 40%, to around 2%. Also removing the red live
dot blink it's reducing it to 0.2% now.
Less sexy UI, but performance is much more important here.