fix(rendering): restore snappiness
Previously app state was triggering observers on every change, so if
there was a change of two fields at once, like offset and cursor, redraw
was triggered twice. Redrawing takes some time, so there was a lag
between changing cursor annd offset. Now I can do a kind of transaction
on appState. If I've put deffer before a change, and commit after, it
will trigger rerender only once. And it works really great. The only
problem here might be, when someone forgot to run commit(), or there
will be concurrent save for some reason, when appState is in batching
mode, and someone is trying to set something. But actually it will be
triggered in a while, so maybe it's not a problem. Will see