refactor(rendering): prepare to render in next tick
Now, after change detection, it will rerender view in the next tick of
main loop. That way it's better to redraw interface not on every change,
and wait until it redraw, then make next change. Just mark that it has
to be redrawed in next tick, so drawing is no longer delaying app state
observing.